JetBrains: Cody: Move classes tied to the language table to a separate package
