+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/ui/CodyConfigurable.kt
@@ -8,12 +8,12 @@ import com.intellij.ui.ColorPanel
 import com.intellij.ui.components.JBCheckBox
 import com.intellij.ui.dsl.builder.*
 import com.intellij.ui.dsl.gridLayout.HorizontalAlign
-import com.sourcegraph.cody.config.AutocompleteLanguageTableWrapper
-import com.sourcegraph.cody.config.AutocompleteLanguageTable
 import com.sourcegraph.cody.config.CodyApplicationSettings
 import com.sourcegraph.cody.config.SettingsModel
 import com.sourcegraph.cody.config.notification.CodySettingChangeActionNotifier
 import com.sourcegraph.cody.config.notification.CodySettingChangeContext
+import com.sourcegraph.cody.config.ui.lang.AutocompleteLanguageTable
+import com.sourcegraph.cody.config.ui.lang.AutocompleteLanguageTableWrapper
 import com.sourcegraph.config.ConfigUtil
 import java.awt.Color
