+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/ui/lang/AutocompleteLanguageTableWrapper.kt
@@ -1,4 +1,4 @@
-package com.sourcegraph.cody.config
+package com.sourcegraph.cody.config.ui.lang
 
 import java.awt.BorderLayout
 import java.awt.Dimension
